Add Enemies and Bosses

+++ Scripts/BitArray.cs
@@ -2,8 +2,11 @@ using System.Diagnostics;
 
 [DebuggerDisplay("{PrintArray(), nq}")]
 public class BitArray {
-	private int array;
+	private uint array;
 
+	/// <summary>
+	/// Initializes the class
+	/// </summary>
 	public BitArray() {
 		array = 0;
 	}
@@ -32,7 +35,7 @@ public class BitArray {
 	/// </summary>
 	/// <param name="bit">The index of the bit to set.</param>
 	/// <param name="on">The value to set the bit to. True for 1, false for 0.</param>
-	public void SetBitVal(int bit, bool on) {
+	public void SetBitVal(uint bit, bool on) {
 		if (on) {
 			SetBit(bit);
 		} else {
@@ -44,7 +47,7 @@ public class BitArray {
 	/// Sets the specified bit in the BitArray.
 	/// </summary>
 	/// <param name="bit">The bit to set.</param>
-	public void SetBit(int bit) {
+	public void SetBit(uint bit) {
 		array |= bit;
 	}
 
@@ -52,10 +55,14 @@ public class BitArray {
 	/// Clears the specified bit in the BitArray.
 	/// </summary>
 	/// <param name="bit">The index of the bit to clear.</param>
-	public void ClearBit(int bit) {
+	public void ClearBit(uint bit) {
 		array &= ~array | bit;
 	}
 
+	/// <summary>
+	/// Returns a string representation of the BitArray
+	/// </summary>
+	/// <returns>The BitArray as a string</returns>
 	public string PrintArray() {
 		string output = "";
 		for (int i = 1; i <= 32; i++) {
@@ -64,11 +71,16 @@ public class BitArray {
 		return output;
 	}
 
+	/// <summary>
+	/// Creates a BitArray from a string representation.
+	/// </summary>
+	/// <param name="stringRep">The string representation of the BitArray.</param>
+	/// <returns>The BitArray created from the string representation.</returns>
 	public static BitArray FromString(string stringRep) {
 		BitArray output = new BitArray();
-		for (int i = 0; i < 32; i++) {
-			if (stringRep[i] == '1') {
-				output.SetBit(2^i);
+		for (uint i = 0; i < 32; i++) {
+			if (stringRep[(int)i] == '1') {
+				output.SetBit(2 ^ i);
 			}
 		}
 		return output;
